+++ pages/_document.tsx
@@ -42,7 +42,6 @@ class MyDocument extends Document {
     return (
       <Html className="overflow-x-hidden">
         <Head>
-          <meta name="viewport" content="width=device-width, initial-scale=1" />
           <meta name="apple-mobile-web-app-capable" content="yes" />
           <meta
             name="apple-mobile-web-app-status-bar-style"
